events/raidWatcher.sh: Fixed a major bug with the event trigger.
When a single raid was in place, the script would not detect if it was
faulty or not simply because the parsing returned an empty string.

The bug itself was fixed and also, a check to see if the parsing
actually returns an empty value was also added to avoid this kind of
problem in the future.