sysWatcher: implemented a new CUI argument system.
Currently, only the configuration file can be set. It is now important
to set the configuration file with the proper flags rather than simply
inputing the configuration path to the command, by itself, like it was
before. The implemented argument system will be used to add more options
and flags to the command in the future.