runScriptFunc: Rewrote the script completely.
This new version of the script now supports loading a virtually infinite
amount of scripts (read libraries) and the added ability to pass values
to the function to call.

This new feature was added to lift the requirement of the event scripts
to include the "global" utils.sh script and also to automatically include the
variables set in the configuration file. In other terms, it is will now
be possible to access functions from util.sh and variables from the
configuration file from the event scripts without the requirement of
including them as they will be automatically included.